refactor(overlay): extract the overlay view layer into its own file (#2292 C1)
Chunk C1 of the #2292 overlay window-ownership ladder. Behaviour-neutral file
move: RecordingOverlayPanel.swift:1976-3368 relocated as ONE unit into
Sources/EnviousWisprAppKit/App/Overlay/Views/OverlayLegacyViews.swift.

It exists so the whole-diff reviewer can dismiss ~1,390 relocated lines in one
read instead of re-reading them as new code when the cutover lands.

The move is byte-identical, measured rather than asserted:
  git show HEAD:...RecordingOverlayPanel.swift | sed -n '1976,3368p'
  vs the new file from its "// MARK: - SpectrumWheelIcon" line to EOF
  -> cmp: identical, 1393 lines
  sed -n '1,1974p' of the original vs the whole new RecordingOverlayPanel.swift
  -> cmp: identical

The fileprivate `extension NSPoint` at :1966-1974 deliberately STAYS. Its
declaration is at :1971 and its only callers are inside the class that survives
until the C4 cutover (:370, :910), so moving it does not compile. That is the
#1341 drag-detection tolerance helper; C4 deletes it with the class.

One file, not one per feature: splitting the view types here would break their
file-private relationships, which is a behaviour-affecting refactor wearing a
file-move's clothes. The per-feature split belongs in C4, where the ownership
change justifies it.

CapsuleBackgroundFreezeTests reads the overlay source as a whole file and is the
only test that does. Retargeted at :48 (the path) and :146 (the diagnostic
filename). The retarget is provably correct rather than assumed: every literal
it freezes was counted on both sides of the split BEFORE the move --
capsule fill head=0 views=2, capsule border head=0 views=1, capsule notice text
head=0 views=1, PreviewPillPalette. head=0 views=10, OverlayCapsuleBackground
head=0 views=1, DistressCapsuleBackground head=0 views=1. Pointing it at the old
file would return 0 for every one and fail loudly.

Three stale cross-file pointers fixed, found by sweeping the old path:
  MenuBarIconAnimator.swift:154, :310 -- cited RainbowLipsIcon and
    SpectrumWheelIcon "in RecordingOverlayPanel.swift"; both moved.
  PreviewPillPalette.swift:14 -- claimed that file "is ~3,000 lines"; it is now
    1,974.
LivePreviewCoordinator.swift:217 and :283 cite :382-385, :640-656 and :375-388,
all below 1974 and therefore unmoved and still accurate. Deliberately left.

Validation: scripts/xcode-test.sh (worktree path) -> ** TEST SUCCEEDED **,
EXIT=0, Debug lane executed 6355 tests, zero failures. Per-bundle 6150 + 205
reconciles exactly with the lane total. Suite CapsuleBackgroundFreezeTests
passed. Zero compile errors; the seven "error:" hits in the log are XPC
connection noise from the test host.

No manifest change needed: Project.swift:234 globs Sources/<name>/** recursively.

Self-audit-grep: /usr/bin/grep -rn 'RecordingOverlayPanel\.swift' Sources Tests
Self-audit-owner: Sources/EnviousWisprAppKit/App/Overlay/Views/OverlayLegacyViews.swift
